hubspot-connector/requirements.txt
pydantic
python-dotenv
hubspot-api-client

pandas>=2.2.1
tabulate